chore(editor): extracts the logic to set back the selection
